Small fixes to USER
