+++ src/main/java/com/praktiker/shop/dto/user/UserRegisterResponse.java
@@ -1,8 +1,6 @@
 package com.praktiker.shop.dto.user;
 
-import com.praktiker.shop.entities.user.Role;
-
 import java.util.Set;
 
-public record UserRegisterResponse(Long id, String username, String email, Set<Role> roles) {
+public record UserRegisterResponse(Long id, String username, String email, Set<String> roles) {
 }
