+++ src/main/java/com/praktiker/shop/mappers/UserMapper.java
@@ -6,16 +6,17 @@ import com.praktiker.shop.entities.user.Role;
 import com.praktiker.shop.entities.user.User;
 
 import java.util.Set;
+import java.util.stream.Collectors;
 
 public class UserMapper {
 
     public static User toEntity(UserRegisterRequest request, String password, Set<Role> roles) {
         return User.builder()
-                .username(request.getUsername())
-                .password(password)
-                .email(request.getEmail())
-                .roles(roles)
-                .build();
+                   .username(request.getUsername())
+                   .password(password)
+                   .email(request.getEmail())
+                   .roles(roles)
+                   .build();
     }
 
     public static UserRegisterResponse toResponse(User user) {
@@ -23,6 +24,6 @@ public class UserMapper {
                 user.getId(),
                 user.getUsername(),
                 user.getEmail(),
-                user.getRoles());
+                user.getRoles().stream().map(Role::getName).collect(Collectors.toSet()));
     }
 }
