+++ src/test/java/com/praktiker/shop/controllers/AuthControllerTest.java
@@ -5,7 +5,7 @@ import com.praktiker.shop.config.security.SecurityConfig;
 import com.praktiker.shop.dto.user.UserRegisterRequest;
 import com.praktiker.shop.dto.user.UserRegisterResponse;
 import com.praktiker.shop.services.AuthService;
-import com.praktiker.shop.utilis.factories.UserRegisterTestFactory;
+import com.praktiker.shop.utilis.factories.UserTestFactory;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
@@ -13,12 +13,13 @@ import org.springframework.context.annotation.Import;
 import org.springframework.http.MediaType;
 import org.springframework.test.context.bean.override.mockito.MockitoBean;
 import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.test.web.servlet.MvcResult;
 
+import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.when;
 import static org.springframework.security.test.web.servlet.request.SecurityMockMvcRequestPostProcessors.csrf;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 @WebMvcTest(AuthController.class)
@@ -36,19 +37,23 @@ public class AuthControllerTest {
 
     @Test
     public void shouldRegisterUser() throws Exception {
-        UserRegisterRequest request = UserRegisterTestFactory.createUserRequest();
+        UserRegisterRequest request = UserTestFactory.createUserRequest();
 
-        UserRegisterResponse response = UserRegisterTestFactory.createUserResponse();
+        UserRegisterResponse response = UserTestFactory.createUserResponse();
 
         when(authService.register(any(UserRegisterRequest.class))).thenReturn(response);
 
-        mockMvc.perform(post("/auth/register")
-                                .with(csrf())
-                                .contentType(MediaType.APPLICATION_JSON)
-                                .content(objectMapper.writeValueAsString(request)))
-               .andExpect(status().isCreated())
-               .andExpect(jsonPath("$.id").value(1L))
-               .andExpect(jsonPath("$.username").value("Adam"))
-               .andExpect(jsonPath("$.email").value("adam@mail.com"));
+        MvcResult mvcResult = mockMvc.perform(post("/auth/register")
+                                                      .with(csrf())
+                                                      .contentType(MediaType.APPLICATION_JSON)
+                                                      .content(objectMapper.writeValueAsString(request)))
+                                     .andExpect(status().isCreated())
+                                     .andReturn();
+
+        String json = mvcResult.getResponse().getContentAsString();
+
+        UserRegisterResponse actual = objectMapper.readValue(json, UserRegisterResponse.class);
+
+        assertEquals(response, actual);
     }
 }
