+++ src/test/java/com/praktiker/shop/utilis/factories/UserTestFactory.java
@@ -1,11 +1,16 @@
 package com.praktiker.shop.utilis.factories;
 
+import com.praktiker.shop.dto.user.UserRegisterRequest;
+import com.praktiker.shop.dto.user.UserRegisterResponse;
 import com.praktiker.shop.entities.user.User;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Set;
 
 public class UserTestFactory {
+
+    /********************************* USER *********************************/
 
     public static List<User> createUsers() {
         List<User> users = new ArrayList<>();
@@ -29,5 +34,29 @@ public class UserTestFactory {
 
     public static User createUser() {
         return createUser("Adam", "Password123", "adam@mail.com");
+    }
+
+    /********************************* USER REGISTER REQUESTS *********************************/
+
+    public static UserRegisterRequest createUserRequest(String username, String password, String email) {
+        return new UserRegisterRequest(username, password, email);
+    }
+
+    public static UserRegisterRequest createUserRequest() {
+        return new UserRegisterRequest("Adam", "Password123", "adam@mail.com");
+    }
+
+    /********************************* USER RESPONSES *********************************/
+
+    public static UserRegisterResponse createUserResponse(Long id, String username, String email, Set<String> roles) {
+        return new UserRegisterResponse(id, username, email, roles);
+    }
+
+    public static UserRegisterResponse createUserResponse() {
+        return new UserRegisterResponse(
+                1L,
+                "Adam",
+                "adam@mail.com",
+                Set.of("ROLE_USER"));
     }
 }
